Split tmux session identity key
